# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Booking Platformasi ucun Backend API. Randevu vaxtlari is qrafiki, nahar fasilesi ve slot addimi esasinda runtime-da hesablanir; bildirisler WebSocket ve FCM ile realtime catdirilir.",
        "title": "Booking Service API",
        "contact": {},
        "version": "2.0"
    },
    "host": "booking-service-sld9.onrender.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/bookings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Biznesin bronlari",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Isci ID",
                        "name": "staff_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Musteri ID",
                        "name": "customer_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pending|confirmed|reschedule_proposed|cancelled|completed|no_show",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 ve ya YYYY-MM-DD",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 ve ya YYYY-MM-DD",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Default 50",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Default 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_booking.SuccessResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Secilen vaxt availability muherrikinden kecirilir. Bos deyilse SLOT_TAKEN qaytarilir. auto_confirm sondurulubse bron pending yaranir ve isciye aninda bildiris gedir.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Bron yarat",
                "parameters": [
                    {
                        "description": "Bron",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_OrkhanNajaf1i_booking-service_internal_domain_booking.CreateBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_booking.SuccessResponse"
                        }
                    },
                    "409": {
                        "description": "Vaxt artiq tutulub",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_booking.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bookings/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Menim bronlarim",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_booking.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/bookings/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Bron detallari",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_booking.SuccessResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Bron qeydlerini yenile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Qeydler",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_OrkhanNajaf1i_booking-service_internal_domain_booking.UpdateBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_booking.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/bookings/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Bronu legv et",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sebeb",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_OrkhanNajaf1i_booking-service_internal_domain_booking.CancelBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_booking.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/bookings/{id}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Randevunu tamamlanmis isarele",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_booking.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/bookings/{id}/confirm": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Bronu tesdiqle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_booking.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/bookings/{id}/no-show": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Musteri gelmedi",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_booking.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/bookings/{id}/propose": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Provider musteriye basqa vaxt teklif edir. Teklif olunan vaxt da qrafik uzre yoxlanilir. Musteri qebul edene qeder ilkin vaxt saxlanilir.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Alternativ vaxt teklif et",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Yeni vaxt",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_OrkhanNajaf1i_booking-service_internal_domain_booking.ProposeRescheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_booking.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/bookings/{id}/respond": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "accept=true olsa bron teklif olunan vaxta kecir ve tesdiqlenir; false olsa ilkin vaxta qayidir.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Teklife cavab ver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cavab",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_OrkhanNajaf1i_booking-service_internal_domain_booking.RespondToProposalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_booking.SuccessResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_OrkhanNajaf1i_booking-service_internal_domain_booking.CancelBookingRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
        "github_com_OrkhanNajaf1i_booking-service_internal_domain_booking.CreateBookingRequest": {
            "type": "object",
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "location_id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "service_id": {
                    "type": "string"
                },
                "staff_id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "github_com_OrkhanNajaf1i_booking-service_internal_domain_booking.ProposeRescheduleRequest": {
            "type": "object",
            "properties": {
                "new_start_time": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                }
            }
        },
        "github_com_OrkhanNajaf1i_booking-service_internal_domain_booking.RespondToProposalRequest": {
            "type": "object",
            "properties": {
                "accept": {
                    "type": "boolean"
                },
                "note": {
                    "type": "string"
                }
            }
        },
        "github_com_OrkhanNajaf1i_booking-service_internal_domain_booking.UpdateBookingRequest": {
            "type": "object",
            "properties": {
                "notes": {
                    "type": "string"
                }
            }
        },
...
        "internal_http_handlers_booking.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "internal_http_handlers_booking.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}